requirements.txt
moviepy==1.0.3
yandexcloud==0.196.0
openai==0.26.5
reprint==0.6.0
